+++ app/models/user.rb
@@ -14,6 +14,7 @@ class User < ApplicationRecord
 
   has_many :users_sessions, class_name: 'Users::Session', dependent: :destroy
   has_many :feedbacks, dependent: :destroy
+  has_many :identities, dependent: :destroy
 
   scope :not_confirmed, -> { where(confirmed_at: nil) }